data mining/CascadeProjects/windsurf-project/hospital_dss/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
pydantic==2.1.1
python-dotenv==1.0.0
scipy==1.10.1
imbalanced-learn==0.10.1
xgboost==1.7.5
pytest==7.3.1
